app/requirements.txt
async-timeout==5.0.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
prometheus_client==0.21.1
redis==5.2.1
Werkzeug==3.1.3
python-dotenv==1.1.0
